Restructure values
